fix: stubs for qsVerifier/Signer
